feat: Implemented hexagonal directories backbone
In this commit I've introduced the structure and an example of use of the Hexagonal architecture
we're going to be using from now on in the project.
The example of case of use with 'Students' shall provide us a template to follow in our developments and will help us immensely as we
just got started with the concept of clean architectures and most importantly how to implement them.