feat: Http client interceptor implemented, student placeholder deleted
I have cleaned the backend functionalities related with students as it was a starting point placeholder
to develop our 1st real functionalities.

In the front end i've implemented a HttpClient Interceptor which will append the header
'Authorization' needed for every request to check the login session and consume the API except
for /auth/login which won't needed and is handled as such.